superdog/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
Pillow==8.3.2
python-gettext==4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
